+++ npkpir_23/src/java/view/StatisticAdminView.java
@@ -68,7 +68,7 @@ public class StatisticAdminView implements Serializable {
                 stat.iloscwydrukow += p.getIloscwydrukow();
                 if (p.getWylogowanie() instanceof Date && p.getZalogowanie() instanceof Date) {
                     Duration duration = new Duration(new DateTime(p.getZalogowanie()),new DateTime(p.getWylogowanie()));
-                    milis += (int) duration.getMillis();
+                    milis += duration.getMillis();
                 }
             }
             int minuty = (int) (milis/1000/60);
